requirements.txt
argcomplete==1.11.1
arrow==0.15.5
beautifulsoup4==4.9.0
boto3==1.12.39
botocore==1.15.39
certifi==2020.4.5.1
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
cssselect==1.1.0
date-guesser==2.1.4
DateTime==4.3
docutils==0.15.2
durationpy==0.5
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.1
jmespath==0.9.5
kappa==0.6.0
langdetect==1.0.8
lxml==4.5.0
MarkupSafe==1.1.1
newspaper3k==0.2.8
nltk==3.4.5
Pillow==7.1.1
pip-tools==4.5.1
placebo==0.9.0
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-file==1.4.3
s3transfer==0.3.3
six==1.14.0
soupsieve==2.0
text-unidecode==1.3
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
tqdm==4.45.0
troposphere==2.6.0
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
zope.interface==5.1.0
